+++ examples/HelloWorldMvc/HelloWorldMvc.fs
@@ -11,14 +11,12 @@ module Model =
     type Greeting =
         { Message : string }
 
-/// Routes templates
 module Route =
     let index = "/"
     let greetPlainText = "/greet/text/{name}"
     let greetJson = "/greet/json/{name}"
     let greetHtml = "/greet/html/{name}"
 
-/// URL factories
 module Url =
     let greetPlainText name = Route.greetPlainText.Replace("{name}", name)
     let greetJson name = Route.greetJson.Replace("{name}", name)
@@ -33,76 +31,70 @@ module View =
             content
 
     module GreetingView =
-        /// HTML view for /greet/html
         let detail greeting =
             layout [
                 Text.h1 $"Hello {greeting.Name} from /html"
                 Elem.hr []
                 Text.p "Greet other ways:"
                 Elem.nav [] [
-                    Elem.a  [ Attr.href (Url.greetPlainText greeting.Name) ] [ Text.raw "Greet in text"]
+                    Elem.a
+                        [ Attr.href (Url.greetPlainText greeting.Name) ]
+                        [ Text.raw "Greet in text"]
                     Text.raw " | "
-                    Elem.a [ Attr.href (Url.greetJson greeting.Name) ] [ Text.raw "Greet in JSON " ]
+                    Elem.a
+                        [ Attr.href (Url.greetJson greeting.Name) ]
+                        [ Text.raw "Greet in JSON " ]
                 ]
             ]
-
-/// Static error page(s)
-module ErrorPage =
-    let notFound : HttpHandler =
-        Response.withStatusCode 404 >>
-        Response.ofHtml (View.layout [ Text.h1 "Not Found" ])
-
-    let serverException : HttpHandler =
-        Response.withStatusCode 500 >>
-        Response.ofHtml (View.layout [ Text.h1 "Server Error" ])
 
 module Controller =
     open Model
     open View
 
+    /// Error page(s)
+    module ErrorController =
+        let notFound : HttpHandler =
+            Response.withStatusCode 404 >>
+            Response.ofHtml (layout [ Text.h1 "Not Found" ])
+
+        let serverException : HttpHandler =
+            Response.withStatusCode 500 >>
+            Response.ofHtml (layout [ Text.h1 "Server Error" ])
+
     module GreetingController =
-        /// GET /
-        let index : HttpHandler =
+        let index =
             Response.ofPlainText "Hello world"
 
-        /// A helper to project the name into an HttpHandler
-        let private mapRouteToNameGreeting next =
-            Request.mapRoute 
-                (fun route -> { Name = route.GetString "name" }) // <-- almost feels like a dynamic 
-                next
+        let plainTextDetail name =
+            Response.ofPlainText $"Hello {name}"
 
-        /// GET /greet/{name}
-        let plainTextDetail : HttpHandler =
-            mapRouteToNameGreeting (fun greeting ->
-            let message = $"Hello {greeting.Name}"
-            Response.ofPlainText message)
+        let jsonDetail name =
+            let message = { Message = $"Hello {name} from /json" }
+            Response.ofJson message
 
-        /// GET /greet/json
-        let jsonDetail : HttpHandler = // <-- Continuation-style HttpHandler
-            mapRouteToNameGreeting (fun greeting ->
-                let message = { Message = $"Hello {greeting.Name} from /json" }
-                Response.ofJson message)
+        let htmlDetail name =
+            { Name = name }
+            |> GreetingView.detail
+            |> Response.ofHtml
 
-        /// GET /greet/html
-        let htmlDetail : HttpHandler =
-            mapRouteToNameGreeting (fun greeting ->
-                greeting
-                |> GreetingView.detail
-                |> Response.ofHtml)
+        let endpoints =
+            let mapRoute (r : RequestData) =
+                r?name.AsString()
+
+            [ get Route.index index
+              mapGet Route.greetPlainText mapRoute plainTextDetail
+              mapGet Route.greetJson mapRoute jsonDetail
+              mapGet Route.greetHtml mapRoute htmlDetail ]
+
+module App =
+    open Controller
+
+    let endpoints =
+        GreetingController.endpoints
 
 module Program =
     open Controller
-    
-    let endpoints = 
-        [ get Route.index GreetingController.index
-          get Route.greetPlainText GreetingController.plainTextDetail
-          get Route.greetJson GreetingController.jsonDetail
-          get Route.greetHtml GreetingController.htmlDetail ]
 
-
-    /// By defining an explicit entry point, we gain access to the command line
-    /// arguments which when passed into Falco are used as the creation arguments
-    /// for the internal WebApplicationBuilder.
     [<EntryPoint>]
     let main args =
         let wapp = WebApplication.Create(args)
@@ -110,10 +102,10 @@ module Program =
         let isDevelopment = wapp.Environment.EnvironmentName = "Development"
 
         wapp.UseIf(isDevelopment, DeveloperExceptionPageExtensions.UseDeveloperExceptionPage)
-            .UseIf(not(isDevelopment), FalcoExtensions.UseFalcoExceptionHandler ErrorPage.serverException)
+            .UseIf(not(isDevelopment), FalcoExtensions.UseFalcoExceptionHandler ErrorController.serverException)
             .Use(StaticFileExtensions.UseStaticFiles)
-            .UseFalco(endpoints)
-            .FalcoNotFound(ErrorPage.notFound)
+            .UseFalco(App.endpoints)
+            .FalcoNotFound(ErrorController.notFound)
             .Run()
 
         0